Updated configurations to handle realm->login-domain->principal hierarchy.
git-svn-id: https://svn.apache.org/repos/asf/geronimo/trunk@328924 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/tomcat-builder/src/test/org/apache/geronimo/tomcat/deployment/TomcatModuleBuilderTest.java
@@ -434,7 +434,7 @@ public class TomcatModuleBuilderTest extends TestCase {
 
         connector = new GBeanData(connectorName, ConnectorGBean.GBEAN_INFO);
         connector.setAttribute("name", "HTTP");
-        connector.setAttribute("port", new Integer(8080));
+        connector.setAttribute("port", new Integer(8181));
         connector.setReferencePattern("TomcatContainer", containerName);
 
         start(container);
